MkdirAllNewAs: error out if dir exists as file
os.MkdirAll() function returns "not a directory" error in case a
directory to be created already exists but is not a directory
(e.g. a file). MkdirAllNewAs function do not replicate the behavior.

This is a bug since it is expected to ensure the required directory
exists and is indeed a directory, and return an error otherwise.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>