MkdirAllNewAs,CopyDirectory: ensure dir perms
Because of umask, sometimes newly created files and directories can have
permission bits different than expected. An easy solution would be to call
os.Chmod after creation -- this is how CopyFile already works.

Now, let's do the same for directories.

PS The alternative to that would be to temporary set umask to 0, but it
is very problematic in Go as it affects other threads.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>